requirements.txt
astroid==2.1.0
autopep8==1.4.3
Babel==2.6.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
entrypoints==0.3
enum34==1.1.6
flake8==3.7.5
future==0.17.1
idna==2.8
IMDbPY==6.6
isodate==0.6.0
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.3.0
mccabe==0.6.1
MechanicalSoup==0.11.0
pep8==1.7.1
Pillow==5.4.1
pycodestyle==2.5.0
pyfiglet==0.8.post1
pyflakes==2.1.0
PyLD==1.0.4
pylint==2.2.2
pyparsing==2.3.1
PyQt5-sip==4.19.14
pytz==2018.9
rdflib==4.2.2
rdflib-jsonld==0.4.0
requests==2.21.0
rotten-tomatoes-client==0.0.3
six==1.12.0
soupsieve==1.7.3
SQLAlchemy==1.2.17
urllib3==1.24.1
wcwidth==0.1.7
Wikidata==0.6.1
wrapt==1.11.1
